requirements.txt
langchain==1.2.10
langchain-core==1.2.15
langchain-text-splitters==1.1.1
langchain-ollama==1.0.1
langchain-chroma==1.1.0
chromadb==1.5.1
fastapi==0.133.0
uvicorn==0.41.0
requests==2.32.5
pydantic==2.12.5
python-dotenv==1.2.1
